refactor(utils): extract gopher parsing into gopher_parse.py
Move parse_gopher_url, parse_menu_line, parse_gopher_menu, sanitize_selector,
format_gopher_url and the item-type category table (_GOPHER_TYPE_CATEGORY /
gopher_type_category) into gopher_parse.py (depends on helpers + models).
utils.py re-exports them. No behavior change.
